pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simpdf"
version = "0.1.0"
description = "Markdown-to-PDF rendering with FPDF2 and configurable TTF fonts."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "Dmitry"}
]
dependencies = [
  "fpdf2>=2.8.0,<3",
  "markdown-it-py>=3.0.0,<4",
]
keywords = ["markdown", "pdf", "fpdf", "cyrillic"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9",
]

[project.scripts]
simpdf = "simpdf.cli:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["simpdf*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
